chore: updated dmt5 calculators carousel
