+++ crowdin/messages.json
@@ -1570,6 +1570,12 @@
   "-1197864059": "Create free demo account",
   "-795459217": "Go to live demo",
   "-2145244263": "This field is required",
+  "-747943809": "Calculate the margin you need to open and hold your positions with our margin calculator.",
+  "-729599084": "DMT5 margin trading calculator",
+  "-1199560191": "Try our margin calculator",
+  "-1419811135": "Calculate your swap fee and know exactly what you are expected to pay or will earn for maintaining an overnight contract.",
+  "-2061765962": "DMT5 swap trading calculator",
+  "-1210185536": "Try our swap calculator",
   "-1360155081": "Take part in the world’s largest financial market where more than $5 trillion worth of currencies are bought and sold each day.",
   "-1235219375": "Trade share price movements of big brands and predict broader market trends with indices that measure the overall performance of a market.",
   "-1959484303": "Cryptocurrencies",
@@ -1857,12 +1863,6 @@
   "-128552689": "Trade forex, stocks, stock indices, and commodities on leverage.",
   "-1542970945": "Trade on the world markets without commission",
   "-507449979": "Trade without commission",
-  "-747943809": "Calculate the margin you need to open and hold your positions with our margin calculator.",
-  "-729599084": "DMT5 margin trading calculator",
-  "-1199560191": "Try our margin calculator",
-  "-1419811135": "Calculate your swap fee and know exactly what you are expected to pay or will earn for maintaining an overnight contract.",
-  "-2061765962": "DMT5 swap trading calculator",
-  "-1210185536": "Try our swap calculator",
   "-1843545572": "With the calculators and numerous analytical tools available on the Deriv MT5 platform, you’ll be able to manage your capital and trading positions better.",
   "-224051432": "24/7",
   "-1080904642": "active traders",
