+++ src/features/dmt5/caclculators-carousel/types.ts
@@ -1,0 +1,11 @@
+import { TString } from 'types/generics'
+
+export type TCalculator = {
+    index: number
+    name: TString
+    text: TString
+    image_name: string
+    image_alt_name: TString
+    button_text: TString
+    link: string
+}
