+++ src/pages/dmt5/_margin_calculator.tsx
@@ -1,71 +1,11 @@
 import React from 'react'
-import Proptypes from 'prop-types'
-import { graphql, useStaticQuery } from 'gatsby'
 import styled from 'styled-components'
-import { Box, Flex, SectionContainer, Desktop, Mobile } from 'components/containers'
-import { Carousel, Header, LinkText, QueryImage, Text } from 'components/elements'
-import { LinkButton } from 'components/form'
-import { Localize, localize } from 'components/localization'
+import { Box, Flex, SectionContainer } from 'components/containers'
+import { Header, LinkText, Text } from 'components/elements'
+import { Localize } from 'components/localization'
 import device from 'themes/device'
 import { useCountryRule } from 'components/hooks/use-country-rule'
-
-type CalculatorProps = {
-    children?: React.ReactNode
-    key?: number
-    index?: number
-    name: React.ReactElement
-    text: React.ReactElement
-    image_name: string
-    image_alt_name: string
-    button_text: React.ReactElement
-    link: string
-}
-
-const query = graphql`
-    query {
-        margin_calculator: file(relativePath: { eq: "dmt5/dmt5-margin-calculator.png" }) {
-            ...fadeIn
-        }
-        margin_calculator_mobile: file(
-            relativePath: { eq: "dmt5/dmt5-margin-calculator-mobile.png" }
-        ) {
-            ...fadeIn
-        }
-        swap_calculator: file(relativePath: { eq: "dmt5/dmt5-swap-calculator.png" }) {
-            ...fadeIn
-        }
-        swap_calculator_mobile: file(relativePath: { eq: "dmt5/dmt5-swap-calculator-mobile.png" }) {
-            ...fadeIn
-        }
-    }
-`
-const CardContainer = styled(Flex)`
-    max-width: 588px;
-    @media ${device.tabletL} {
-        max-width: 95%;
-    }
-`
-
-const ImageWrapper = styled.div`
-    display: flex;
-    width: 392px;
-    height: 386px;
-    object-fit: contain;
-    margin-bottom: 2.4rem;
-
-    @media ${device.tabletL} {
-        align-items: center;
-        justify-content: center;
-        width: 232px;
-        height: 229px;
-
-        div {
-            max-width: 232px;
-            width: 100%;
-            height: auto;
-        }
-    }
-`
+import CalculatorsCarousel from 'features/dmt5/caclculators-carousel'
 
 const MainHeader = styled(Header)`
     margin: 0 0 12px;
@@ -74,15 +14,6 @@ const MainHeader = styled(Header)`
         margin-bottom: 24px;
         padding: 0 16px;
         text-align: center;
-    }
-`
-
-const SubHeader = styled(Header)`
-    font-size: 24px;
-    color: var(--color-black-3);
-    @media ${device.tabletL} {
-        font-size: 20px;
-        margin-bottom: 8px;
     }
 `
 
@@ -99,46 +30,6 @@ const StyledText = styled(Text)`
     margin-bottom: 24px;
 `
 
-const CardText = styled(StyledText)`
-    margin: 0.8rem 4rem 2.4rem;
-    @media ${device.tabletL} {
-        margin: 0 0 24px;
-    }
-`
-
-const StyledCardContainer = styled(Flex)`
-    width: 100%;
-    height: 684px;
-    flex-direction: column;
-    flex-wrap: wrap;
-    justify-content: center;
-    align-items: center;
-    @media ${device.tabletL} {
-        height: auto;
-        min-height: 480px;
-        justify-content: flex-start;
-    }
-`
-
-const StyledLinkButton = styled(LinkButton)`
-    padding: 1rem 1.6rem;
-    font-size: 14px;
-    height: 40px;
-    width: auto;
-    min-width: 210px;
-    @media ${device.tabletL} {
-        padding: 10px 15px;
-    }
-
-    :active {
-        outline: none;
-        border: none;
-    }
-    :focus {
-        outline: 0;
-    }
-`
-
 const StyledFlexContainer = styled(Flex)`
     width: 100%;
     flex-wrap: nowrap;
@@ -149,7 +40,7 @@ const StyledFlexContainer = styled(Flex)`
     }
 `
 
-const StyledFlex = styled(Flex)`
+const StyledFlex = styled(Flex)<{ has_color?: boolean }>`
     width: 50%;
     min-height: 694px;
     margin-right: 2.4rem;
@@ -161,90 +52,7 @@ const StyledFlex = styled(Flex)`
     }
 `
 
-const CalculatorCard = ({
-    button_text,
-    image_alt_name = '',
-    image_name = '',
-    link = '',
-    name,
-    text,
-}: CalculatorProps) => {
-    const data = useStaticQuery(query)
-    return (
-        <StyledCardContainer>
-            <SubHeader as="h3" align="center">
-                {name}
-            </SubHeader>
-            <CardText align="center">{text}</CardText>
-            <ImageWrapper>
-                <Desktop>
-                    <QueryImage data={data[image_name]} alt={image_alt_name} />
-                </Desktop>
-                <Mobile>
-                    <QueryImage data={data[image_name + '_mobile']} alt={image_alt_name} />
-                </Mobile>
-            </ImageWrapper>
-            <StyledLinkButton tertiary to={link}>
-                <Localize translate_text={button_text} />
-            </StyledLinkButton>
-        </StyledCardContainer>
-    )
-}
-
-CalculatorCard.propTypes = {
-    button_text: Proptypes.string,
-    image_alt_name: Proptypes.string,
-    image_name: Proptypes.string,
-    is_mobile: Proptypes.bool,
-    link: Proptypes.string,
-    name: Proptypes.string,
-    text: Proptypes.string,
-}
-
-const calculators: CalculatorProps[] = [
-    {
-        index: 0,
-        name: <Localize translate_text="Margin calculator" />,
-        text: (
-            <Localize translate_text="Calculate the margin you need to open and hold your positions with our margin calculator." />
-        ),
-        image_name: 'margin_calculator',
-        image_alt_name: localize('DMT5 margin trading calculator'),
-        button_text: <Localize translate_text="Try our margin calculator" />,
-        link: '/trader-tools/margin-calculator/',
-    },
-    {
-        index: 1,
-        name: <Localize translate_text="Swap calculator" />,
-        text: (
-            <Localize translate_text="Calculate your swap fee and know exactly what you are expected to pay or will earn for maintaining an overnight contract." />
-        ),
-        image_name: 'swap_calculator',
-        image_alt_name: localize('DMT5 swap trading calculator'),
-        button_text: <Localize translate_text="Try our swap calculator" />,
-        link: '/trader-tools/swap-calculator/',
-    },
-]
-
 const MarginCalculator = () => {
-    const settings = {
-        container_style: {
-            maxWidth: '100%',
-            margin: '0',
-        },
-        slide_style: {
-            minWidth: '100%',
-            position: 'relative',
-        },
-        chevron_style: {
-            chevron_color: 'red',
-            is_displayed_on_mobile: true,
-        },
-        navigation_style: {
-            nav_color: 'red',
-        },
-    }
-
     const { is_uk_eu } = useCountryRule()
     return (
         <SectionContainer>
@@ -302,27 +110,13 @@ const MarginCalculator = () => {
                 </StyledFlex>
                 <StyledFlex
                     fd="row"
-                    jc="flex-start"
+                    jc="center"
                     tablet_jc="center"
                     wrap="wrap"
                     ml="2.4rem"
                     tabletL={{ ml: '0px', pt: '24px', pl: '16px', pr: '16px' }}
                 >
-                    <CardContainer>
-                        <Carousel {...settings}>
-                            {calculators.map((calculator, idx) => (
-                                <CalculatorCard
-                                    key={idx}
-                                    name={calculator.name}
-                                    image_name={calculator.image_name}
-                                    image_alt_name={calculator.image_alt_name}
-                                    text={calculator.text}
-                                    link={calculator.link}
-                                    button_text={calculator.button_text}
-                                />
-                            ))}
-                        </Carousel>
-                    </CardContainer>
+                    <CalculatorsCarousel />
                 </StyledFlex>
             </StyledFlexContainer>
         </SectionContainer>
